refactor: Refactors code in compare_schemas.py Splits the code in compares_schemas.py into the following modules: - diff.py - Tracks diffs in properties or attributes - version.py - Manages logic around schema version numbers - release.py - Manages the creation of a new release
